Refactored the Market into Presentation and  Container Components
